Payout db method

+++ model/payout.go
@@ -4,7 +4,7 @@ import "time"
 
 type Payout struct {
 	TxID       string    `json:"tx_id,omitempty"`
-	Amount     uint64    `json:"amount,omitempty"`
+	Amount     float64   `json:"amount,omitempty"`
 	Address    string    `json:"address,omitempty"`
 	IsMined    bool      `json:"is_mined,omitempty"`
 	InsertedAt time.Time `json:"inserted_at,omitempty"`
