+++ res/migrations/20201011133517_load_payout_table.up.sql
@@ -1,7 +1,7 @@
 create table payout (
   txid varchar primary key,
   address varchar not null,
-  amount int not null,
+  amount float not null,
   is_mined boolean not null,
   inserted_at timestamp,
   updated_at timestamp
